requirements.txt
PyQt6>=6.6.0
python-vlc>=3.0.0
mutagen>=1.47.0
